# docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "version": "2.0.0",
    "title": "Volcanoes around the world v2",
    "description": "This API has been created to support assignment three in CAB230, Web Computing. It exposes a number of REST endpoints which implement CRUD operations on a database containing publicly available data about volcanoes collated by the Smithsonian Institution's Global Volcanism Program. The API endpoints and their usage are described in detail below."
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Enter JWT Bearer token **_only_**"
      }
    },
...
  },
  "paths": {
...
    "/comment/{volcanoID}": {
      "post": {
        "tags": [
          "Comments"
        ],
        "parameters": [
          {
            "name": "volcanoID",
            "in": "path",
            "description": "The ID of the volcano",
            "required": true,
            "schema": {
              "type": "number"
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "A route to post a comment with a volcano ID.",
        "responses": {
          "201": {
            "description": "An object containing your name and student number.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "boolean",
                      "example": "false"
                    },
                    "message": {
                      "type": "string",
                      "example": "Comment posted"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input: userID and comment are required, and comment must be a non-empty string",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "boolean",
                      "example": "true"
                    },
                    "message": {
                      "type": "string",
                      "example": "Invalid input: userID and comment are required, and comment must be a non-empty string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized. You are required to sign in to comment.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "boolean",
                      "example": "true"
                    },
                    "message": {
                      "type": "string",
                      "example": "Unauthorized. You are required to sign in to comment."
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Volcano not found..",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "boolean",
                      "example": "true"
                    },
                    "message": {
                      "type": "string",
                      "example": "Volcano not found."
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "User not found.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "boolean",
                      "example": "true"
                    },
                    "message": {
                      "type": "string",
                      "example": "User not found."
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "boolean",
                      "example": "true"
                    },
                    "message": {
                      "type": "string",
                      "example": "Internal Server Error"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/comment/{commentID}": {
      "get": {
        "tags": [
          "Comments"
        ],
        "parameters": [
          {
            "name": "commentID",
            "in": "path",
            "description": "The ID of the comment",
            "required": true,
            "schema": {
              "type": "number"
            }
          }
        ],
        "description": "A route to retrieve a comment using volcano ID",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string",
                  "example": "mike@gmail.com"
                },
                "password": {
                  "type": "string",
                  "example": "password"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "An object containing your name and student number.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "commentID": {
                      "type": "string",
                      "example": "3"
                    },
                    "volcanoID": {
                      "type": "string",
                      "example": "1"
                    },
                    "userID": {
                      "type": "string",
                      "example": "1"
                    },
                    "timeStamp": {
                      "type": "string",
                      "example": "2024-06-02T01:14:58.000Z"
                    },
                    "comment": {
                      "type": "string",
                      "example": "CAB230"
                    },
                    "likes": {
                      "type": "string",
                      "example": "2"
                    },
                    "dislikes": {
                      "type": "string",
                      "example": "1"
                    },
                    "ratio": {
                      "type": "string",
                      "example": "66.67%"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input: commentID required. CommentID as an int.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "boolean",
                      "example": "true"
                    },
                    "message": {
                      "type": "string",
                      "example": "Invalid input: commentID required. CommentID as an int."
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Comment does not exist.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "boolean",
                      "example": "true"
                    },
                    "message": {
                      "type": "string",
                      "example": "Comment does not exist."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/comment/{commentID}/{likeOrDislike}": {
      "put": {
        "tags": [
          "Comments"
        ],
        "parameters": [
          {
            "name": "commentID",
            "in": "path",
            "description": "The ID of the comment",
            "required": true,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "likeOrDislike",
            "in": "path",
            "description": "like or dislike the comment",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "like",
                "dislike"
              ]
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "A route to either like or dislike a comment",
        "responses": {
          "201": {
            "description": "Sucess message when a comment is liked or disliked.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "boolean",
                      "example": "false"
                    },
                    "message": {
                      "type": "string",
                      "example": "Comment liked or disliked"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input: commentID, and like or dislike required. CommentID as an int, like or dislike as string.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "boolean",
                      "example": "true"
                    },
                    "message": {
                      "type": "string",
                      "example": "Invalid input: commentID, and like or dislike required. CommentID as an int, like or dislike as string."
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Comment does not exist.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "boolean",
                      "example": "true"
                    },
                    "message": {
                      "type": "string",
                      "example": "Comment does not exist."
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "boolean",
                      "example": "true"
                    },
                    "message": {
                      "type": "string",
                      "example": "Internal Server Error"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}